Rename client to perfclient
